% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcautoarima.R
\name{bcautoarima}
\alias{bcautoarima}
\title{bcautoarima}
\usage{
bcautoarima(
  x,
  trainratio,
  validationratio = 0,
  testratio,
  stationaryspi,
  spiscale,
  ...
)
}
\arguments{
\item{x}{data.table}

\item{trainratio}{Numeric value represents the proportion of the training set}

\item{validationratio}{Numeric value represents the proportion of the validation set}

\item{testratio}{Numeric value represents the proportion of the test set}

\item{stationaryspi}{Logical when TRUE SPI is calculated; when FALSE NSPI is calculated}

\item{spiscale}{Numeric value that reflects the scale of the index}

\item{...}{Additional arguments that relate to the inputs of forecast::auto.arima}
}
\value{
list with evaluation metrics and diagnostic plots
}
\description{
Fits a bias-corrected auto.arima to forecast the standardized precipitation index
}
\examples{
x = dummyrainfall(start = 1950, end = 2020)
bcautoarima(x, 0.8, 0, 0.2, TRUE, 12, seasonal = FALSE)
}
