% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcautoarimapred.R
\name{bcautoarimapred}
\alias{bcautoarimapred}
\title{bcautoarimapred}
\usage{
bcautoarimapred(x, model, stationaryspi, spiscale, timesteps, ...)
}
\arguments{
\item{x}{A data.table with two columns (Date, Rainfall)}

\item{model}{If NULL a new model is built, else a forecast::Arima model object is trained}

\item{stationaryspi}{Logical when TRUE SPI is calculated; when FALSE NSPI is calculated}

\item{spiscale}{Numeric value that reflects the scale of the index}

\item{timesteps}{Number of periods for forecasting}

\item{...}{Additional arguments that relate to the inputs of forecast::auto.arima}
}
\value{
numeric value with model predictions
}
\description{
Fits or updates a forecast::autoarima model and predicts the next observation(s)
}
\examples{
x = dummyrainfall(start = 1950, end = 2020)
bcautoarimapred(x, NULL, TRUE, 12, 1, seasonal = FALSE)
}
