% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmodwtautoarima.R
\name{bcmodwtautoarima}
\alias{bcmodwtautoarima}
\title{bcmodwtautoarima}
\usage{
bcmodwtautoarima(
  x,
  trainratio,
  validationratio = 0,
  testratio,
  stationaryspi,
  spiscale,
  modwtfilter,
  nlevels,
  ...
)
}
\arguments{
\item{x}{data.table}

\item{trainratio}{Numeric value represents the proportion of the training set}

\item{validationratio}{Numeric value represents the proportion of the validation set}

\item{testratio}{Numeric value represents the proportion of the test set}

\item{stationaryspi}{Logical when TRUE SPI is calculated; when FALSE NSPI is calculated}

\item{spiscale}{Numeric value that reflects the scale of the index}

\item{...}{Additional arguments that relate to the inputs of forecast::auto.arima}
}
\value{
list with evaluation metrics and diagnostic plots
}
\description{
Fits a bias-corrected auto.arima with Maximal Overlap Discrete Wavelet Transformation (MODWT) to forecast the standardized precipitation index
}
\examples{
x = dummyrainfall(start = 1950, end = 2020)
bcmodwtautoarima(x, 0.8, 0, 0.2, TRUE, 12, 'haar', 2)
}
