% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcoosautoarima.R
\name{bcoosautoarima}
\alias{bcoosautoarima}
\title{bcoosautoarima}
\usage{
bcoosautoarima(
  x,
  trainratio,
  validationratio = 0,
  testratio,
  stationaryspi,
  spiscale,
  ...
)
}
\arguments{
\item{x}{data.table}

\item{trainratio}{numeric value represents the proportion of the training set}

\item{validationratio}{numeric value represents the proportion of the validation set}

\item{testratio}{numeric value represents the proportion of the test set}

\item{stationaryspi}{logical when TRUE SPI is calculated; when FALSE NSPI is calculated}

\item{spiscale}{numeric the scale of the index}

\item{...}{construct that relates to the inputs of forecast::auto.arima}
}
\value{
list with evaluation metrics and diagnostic plots
}
\description{
bcoosautoarima
}
\examples{
x= dummyrainfall(start = 1950, end = 2020)
bcoosautoarima(x, 0.8, 0, 0.2, TRUE, 12, seasonal = FALSE)
}
