% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compdist.R
\name{compdist}
\alias{compdist}
\title{compdist}
\usage{
compdist(x, mDist, stationaryspi, spiScale)
}
\arguments{
\item{x}{\code{data.table} with monthly rainfall data (Date | Rainfall)}

\item{mDist}{An character vector of one or more}

\item{stationaryspi}{logical used in \emph{array} method (TRUE when stationary and FALSE when otherwise)}

\item{spiScale}{A positive integer value, typically ranging from 1 to 24}
}
\value{
data.table
}
\description{
Compares drought indices fitted with different distributions
}
\examples{
compdist(x = dummyrainfall(1950, 2000),
         mDist = 'gamma',
         stationaryspi = TRUE,
         spiScale = 12
         )
}
