% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeclass.R
\name{computeclass}
\alias{computeclass}
\title{computeclass}
\usage{
computeclass(object)
}
\arguments{
\item{object}{drought class object}
}
\value{
data.table
}
\description{
Computes the drought class of a meteorological drought index
}
\examples{
rainfall = dummyrainfall(1950, 2000)
droughtindex = computenspi(x = dummyrainfall(1950, 2000),
                           stationaryspi = TRUE,
                           spiScale = 12,
                           dist = 'gamma'
                          )
computeclass(droughtindex)
}
