% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computenspi.R
\name{computenspi}
\alias{computenspi}
\title{computenspi}
\usage{
computenspi(x, stationaryspi, spiScale, dist = "gamma")
}
\arguments{
\item{x}{data.table}

\item{stationaryspi}{logical used in \emph{array} method (TRUE when stationary and FALSE when otherwise)}

\item{spiScale}{A positive integer value, typically ranging from 1 to 24}

\item{dist}{A character that can be either, "normal" when Normal distribution is used, "gamma" for Gamma, "zigamma" for zero adjusted Gamma and "weibull" for Weibull.}
}
\value{
data.table
}
\description{
Computes the stationary and non-stationary version of the Standardized Precipitation Index.
The non-stationary version uses GAMLSS and models the parameters of a Gamma distribution
by incorporating the trend of accumulated precipitation.
}
\examples{
computenspi(x = dummyrainfall(1950, 2000),
            stationaryspi = TRUE,
            spiScale = 12,
            dist = 'gamma'
            )
}
