% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droughtclass.R
\name{droughtclass}
\alias{droughtclass}
\title{droughtclass}
\usage{
droughtclass(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Assigns classes to various ranges of the standardized precipitation index (SPI)
}
\examples{
droughtclass(1.5)
}
