% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyrainfall.R
\name{dummyrainfall}
\alias{dummyrainfall}
\title{Generate synthetic monthly rainfall data}
\usage{
dummyrainfall(startYear, endYear)
}
\arguments{
\item{startYear}{A numeric positive integer indicating the start year (e.g., 1950)}

\item{endYear}{A numeric numeric positive integer indicating the end year (e.g., 2021)}
}
\value{
data.table
}
\description{
Create a data.table with synthetic rainfall data generated from a Gamma distribution
}
\examples{
dummyrainfall(1950, 2021)
}
