% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurebias.R
\name{measurebias}
\alias{measurebias}
\title{measurebias}
\usage{
measurebias(x, trainratio, validationratio, testratio, stationaryspi, spiscale)
}
\arguments{
\item{x}{data.table with rainfall records (Date, Rainfall)}

\item{trainratio}{numeric; the proportion of the train set}

\item{validationratio}{numeric; the proportion of the validation set}

\item{testratio}{numeric; the proportion of the test set}

\item{stationaryspi}{logical; When TRUE SPI is calculated; when FALSE NSPI is calculated}

\item{spiscale}{integer; the scale of accumulated precipitation}
}
\value{
list
}
\description{
Measures the amount of bias introduced to the training set due to incorrect computation of (N)SPI prior to out-of-sample validation
}
\examples{
rainfall = dummyrainfall(1950, 2000)
measurebias(rainfall, 0.6, 0.2, 0.2, TRUE, 12)
}
