% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oossplit.R
\name{oossplit}
\alias{oossplit}
\title{oossplit}
\usage{
oossplit(x, trainratio, validationratio, testratio)
}
\arguments{
\item{x}{data.table}

\item{trainratio}{numeric}

\item{validationratio}{numeric}

\item{testratio}{numeric}
}
\value{
data.table
}
\description{
Out-of-sample split: Splits a data.table of class drought into train validation and test sets by preserving chronological order
}
\examples{
rainfall = dummyrainfall(1950, 2000)
rainfall = oossplit(x = rainfall, trainratio = 0.6, validationratio = 0.2, testratio = 0.2)
}
