% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.drought}
\alias{plot.drought}
\title{plot.drought}
\usage{
\method{plot}{drought}(x, label = TRUE, log = TRUE, type = "drought", ...)
}
\arguments{
\item{x}{data.table of class drought}

\item{label}{logical}

\item{log}{logical}

\item{type}{character}

\item{...}{some arguments}
}
\value{
ggplot2
}
\description{
Plots spi or nspi index of a data.table input with class drought
}
\examples{
mIndex = computenspi(x = dummyrainfall(1950, 2000),
                stationary = TRUE,
                spiScale = 12,
                dist = 'gamma')
plot(mIndex)
}
