% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.droughtclass.R
\name{plot.droughtclass}
\alias{plot.droughtclass}
\title{plot.droughtclass}
\usage{
\method{plot}{droughtclass}(x, label = TRUE, log = TRUE, type = "droughtclass", ...)
}
\arguments{
\item{x}{data.table of class droughtclass}

\item{label}{logical}

\item{log}{logical}

\item{type}{character}

\item{...}{some arguments}
}
\value{
ggplot2
}
\description{
Plots spi or nspi index of a data.table input with class drought
}
\examples{
mIndex = computenspi(x = dummyrainfall(1950, 2000),
                stationary = TRUE,
                spiScale = 12,
                dist = 'gamma')

                indexClass = computeclass(mIndex)
plot(indexClass)
}
