% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computenspi.R
\name{computenspi}
\alias{computenspi}
\title{computenspi}
\usage{
computenspi(monthlyRainfall, stationaryspi, spiScale)
}
\arguments{
\item{monthlyRainfall}{data.table}

\item{stationaryspi}{logical}

\item{spiScale}{numeric}
}
\value{
data.table
}
\description{
Computes the stationary and non-stationary version of the Standardized Precipitation Index.
The non-stationary version uses GAMLSS and models the parameters of a Gamma distribution
by incorporating the trend of accumulated precipitation.
}
\examples{
computenspi(monthlyRainfall = dummyrainfall(1950, 2000), stationaryspi = TRUE, spiScale = 12)
}
