% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyrainfall.R
\name{dummyrainfall}
\alias{dummyrainfall}
\title{Generate synthetic monthly rainfall data}
\usage{
dummyrainfall(startYear, endYear)
}
\arguments{
\item{startYear}{numeric}

\item{endYear}{numeric}
}
\value{
data.table
}
\description{
Create a data.table with synthetic rainfall data generated from a Gamma distribution
}
\examples{
dummyrainfall(1950, 2021)
}
